[BE] feat: 전체 경로 탐색 알고리즘 구현 (#38)
- 변경 내용 1 : 전체 경로 탐색 알고리즘 구현 - 외판원 순회 알고리즘 활용

+++ BE/taiso/src/main/java/com/d212/taiso/domain/route/dto/LocationPayload.java
@@ -16,7 +16,7 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class LocationPayload {
 
-    private List<Long> distanceList;
+    private List<Integer> distanceList;
     private long placeId;
     private long rsvId;
 }
